# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KenyaPay Payroll System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - in: query
          name: departmentId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [active, terminated]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/payslips:
    get:
      operationId: getEmployeePayslips
      tags: [employees]
      summary: Get payslips for a specific employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payslips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollEntry"
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /me:
    get:
      operationId: getMe
      tags: [employees]
      summary: Get current user's employee profile
      responses:
        "200":
          description: Current employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
components:
  schemas:
...
    Employee:
      type: object
      required:
        - id
        - employeeNumber
        - firstName
        - lastName
        - email
        - grossSalary
        - employmentType
        - paymentMethod
        - status
        - hireDate
      properties:
        id:
          type: integer
        employeeNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        grossSalary:
          type: number
        employmentType:
          type: string
          enum: [permanent, contract]
        paymentMethod:
          type: string
          enum: [mpesa, bank]
        mpesaNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccount:
          type: ["string", "null"]
        bankBranch:
          type: ["string", "null"]
        kraPin:
          type: ["string", "null"]
        nssfNumber:
          type: ["string", "null"]
        shifNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, terminated]
        hireDate:
          type: string
        terminationDate:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, hr, employee]
        clerkUserId:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required:
        - firstName
        - lastName
        - email
        - grossSalary
        - employmentType
        - paymentMethod
        - hireDate
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        departmentId:
          type: integer
        jobTitle:
          type: string
        grossSalary:
          type: number
        employmentType:
          type: string
          enum: [permanent, contract]
        paymentMethod:
          type: string
          enum: [mpesa, bank]
        mpesaNumber:
          type: string
        bankName:
          type: string
        bankAccount:
          type: string
        bankBranch:
          type: string
        kraPin:
          type: string
        nssfNumber:
          type: string
        shifNumber:
          type: string
        hireDate:
          type: string
        role:
          type: string
          enum: [admin, hr, employee]
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        departmentId:
          type: ["integer", "null"]
        jobTitle:
          type: string
        grossSalary:
          type: number
        employmentType:
          type: string
          enum: [permanent, contract]
        paymentMethod:
          type: string
          enum: [mpesa, bank]
        mpesaNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccount:
          type: ["string", "null"]
        bankBranch:
          type: ["string", "null"]
        kraPin:
          type: string
        nssfNumber:
          type: string
        shifNumber:
          type: string
        status:
          type: string
          enum: [active, terminated]
        role:
          type: string
          enum: [admin, hr, employee]
        terminationDate:
          type: ["string", "null"]
    Department:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        employeeCount:
          type: integer
    DepartmentInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    PayrollEntry:
      type: object
      required:
        - id
        - payrollRunId
        - employeeId
        - grossSalary
        - netPay
      properties:
        id:
          type: integer
        payrollRunId:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeNumber:
          type: string
        departmentName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        grossSalary:
          type: number
        basicSalary:
          type: number
        paye:
          type: number
        nssfEmployee:
          type: number
        nssfEmployer:
          type: number
        shif:
          type: number
        housingLevyEmployee:
          type: number
        housingLevyEmployer:
          type: number
        personalRelief:
          type: number
        insuranceRelief:
          type: number
        otherDeductions:
          type: number
        netPay:
          type: number
        paymentMethod:
          type: string
        disbursementStatus:
          type: string
          enum: [pending, processing, success, failed]
        month:
          type: integer
        year:
          type: integer
...