# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KenyaPay Payroll System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeEmploymentType } from './employeeEmploymentType';
import type { EmployeePaymentMethod } from './employeePaymentMethod';
import type { EmployeeRole } from './employeeRole';
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: number;
  employeeNumber: string;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  grossSalary: number;
  employmentType: EmployeeEmploymentType;
  paymentMethod: EmployeePaymentMethod;
  /** @nullable */
  mpesaNumber?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccount?: string | null;
  /** @nullable */
  bankBranch?: string | null;
  /** @nullable */
  kraPin?: string | null;
  /** @nullable */
  nssfNumber?: string | null;
  /** @nullable */
  shifNumber?: string | null;
  status: EmployeeStatus;
  hireDate: string;
  /** @nullable */
  terminationDate?: string | null;
  role?: EmployeeRole;
  /** @nullable */
  clerkUserId?: string | null;
  createdAt?: string;
}